refactor: move commands to separate files
