+++ Cargo.toml
@@ -24,7 +24,7 @@ pacquet_cafs         = { path = "crates/cafs" }
 
 # Dependencies
 async-trait        = { version = "0.1.72" }
-clap               = { version = "4", features = ["derive"] }
+clap               = { version = "4", features = ["derive", "string"] }
 home               = { version = "0.5.5" }
 insta              = { version = "1.31.0", features = ["yaml", "glob", "walkdir"] }
 futures-util       = { version = "0.3.28" }
