+++ crates/cli/src/commands/mod.rs
@@ -1,10 +1,18 @@
 pub mod add;
 pub mod install;
+pub mod run;
+pub mod store;
 
-use std::path::PathBuf;
+use std::{env, ffi::OsString, path::PathBuf};
 
+use crate::commands::{
+    add::AddCommandArgs, install::InstallCommandArgs, run::RunCommandArgs, store::StoreSubcommands,
+};
 use clap::{Parser, Subcommand};
-use pacquet_package_json::DependencyGroup;
+
+fn default_current_dir() -> OsString {
+    env::current_dir().expect("failed to get current directory").into_os_string()
+}
 
 /// Experimental package manager for node.js written in rust.
 #[derive(Parser, Debug)]
@@ -17,8 +25,8 @@ pub struct Cli {
     pub subcommand: Subcommands,
 
     /// Run as if pacquet was started in <path> instead of the current working directory.
-    #[arg(short = 'C', long = "dir")]
-    pub current_dir: Option<PathBuf>,
+    #[arg(short = 'C', long = "dir", default_value = default_current_dir())]
+    pub current_dir: PathBuf,
 }
 
 #[derive(Subcommand, Debug)]
@@ -26,99 +34,16 @@ pub enum Subcommands {
     /// Initialize a package.json
     Init,
     /// Add a package
-    Add(AddArgs),
+    Add(AddCommandArgs),
     /// Install packages
-    Install(InstallArgs),
+    Install(InstallCommandArgs),
     /// Runs a package's "test" script, if one was provided.
     Test,
     /// Runs a defined package script.
-    Run(RunArgs),
+    Run(RunCommandArgs),
     /// Runs an arbitrary command specified in the package's start property of its scripts object.
     Start,
     /// Managing the package store.
     #[clap(subcommand)]
     Store(StoreSubcommands),
-}
-
-#[derive(Parser, Debug)]
-pub struct AddArgs {
-    /// Name of the package
-    pub package: String,
-    /// Install the specified packages as regular dependencies.
-    #[arg(short = 'P', long = "save-prod", group = "dependency_group")]
-    save_prod: bool,
-    /// Install the specified packages as devDependencies.
-    #[arg(short = 'D', long = "save-dev", group = "dependency_group")]
-    save_dev: bool,
-    /// Install the specified packages as optionalDependencies.
-    #[arg(short = 'O', long = "save-optional", group = "dependency_group")]
-    save_optional: bool,
-    /// Saved dependencies will be configured with an exact version rather than using
-    /// pacquet's default semver range operator.
-    #[arg(short = 'E', long = "save-exact")]
-    pub save_exact: bool,
-    /// The directory with links to the store (default is node_modules/.pacquet).
-    /// All direct and indirect dependencies of the project are linked into this directory
-    #[arg(long = "virtual-store-dir", default_value = "node_modules/.pacquet")]
-    pub virtual_store_dir: String,
-}
-
-#[derive(Parser, Debug)]
-pub struct InstallArgs {
-    /// pacquet will not install any package listed in devDependencies and will remove those insofar
-    /// they were already installed, if the NODE_ENV environment variable is set to production.
-    /// Use this flag to instruct pacquet to ignore NODE_ENV and take its production status from this
-    /// flag instead.
-    #[arg(short = 'P', long = "prod")]
-    pub prod: bool,
-    /// Only devDependencies are installed and dependencies are removed insofar they were
-    /// already installed, regardless of the NODE_ENV.
-    #[arg(short = 'D', long = "dev")]
-    pub dev: bool,
-    /// optionalDependencies are not installed.
-    #[arg(long = "no-optional")]
-    pub no_optional: bool,
-}
-
-impl AddArgs {
-    pub fn get_dependency_group(&self) -> DependencyGroup {
-        if self.save_dev {
-            DependencyGroup::Dev
-        } else if self.save_optional {
-            DependencyGroup::Optional
-        } else {
-            DependencyGroup::Default
-        }
-    }
-}
-
-#[derive(Parser, Debug)]
-pub struct RunArgs {
-    /// A pre-defined package script.
-    pub command: String,
-
-    /// Any additional arguments passed after the script name
-    pub args: Vec<String>,
-
-    /// You can use the --if-present flag to avoid exiting with a non-zero exit code when the
-    /// script is undefined. This lets you run potentially undefined scripts without breaking the
-    /// execution chain.
-    #[arg(long = "if-present")]
-    pub if_present: bool,
-}
-
-#[derive(Subcommand, Debug)]
-pub enum StoreSubcommands {
-    /// Checks for modified packages in the store.
-    Store,
-    /// Functionally equivalent to pnpm add, except this adds new packages to the store directly
-    /// without modifying any projects or files outside of the store.
-    Add,
-    /// Removes unreferenced packages from the store.
-    /// Unreferenced packages are packages that are not used by any projects on the system.
-    /// Packages can become unreferenced after most installation operations, for instance when
-    /// dependencies are made redundant.
-    Prune,
-    /// Returns the path to the active store directory.
-    Path,
 }
